增加类的静态代理支持
