+++ library/think/Hook.php
@@ -106,7 +106,7 @@ class Hook
      */
     public static function exec($class, $tag = '', &$params = null, $extra = null)
     {
-        App::$debug && Debug::remark('behavior_start', 'time');
+        App::isDebug() && Debug::remark('behavior_start', 'time');
         $method = Loader::parseName($tag, 1, false);
         if ($class instanceof \Closure) {
             $result = call_user_func_array($class, [ & $params, $extra]);
@@ -126,7 +126,7 @@ class Hook
             $method = ($tag && is_callable([$obj, $method])) ? $method : 'run';
             $result = $obj->$method($params, $extra);
         }
-        if (App::$debug) {
+        if (App::isDebug()) {
             Debug::remark('behavior_end', 'time');
             Log::record('[ BEHAVIOR ] Run ' . $class . ' @' . $tag . ' [ RunTime:' . Debug::getRangeTime('behavior_start', 'behavior_end') . 's ]', 'info');
         }
