+++ library/think/cache/driver/Memcached.php
@@ -86,6 +86,7 @@ class Memcached extends Driver
      */
     public function get($name, $default = false)
     {
+        $this->readTimes++;
         $result = $this->handler->get($this->getCacheKey($name));
         return false !== $result ? $result : $default;
     }
@@ -100,6 +101,7 @@ class Memcached extends Driver
      */
     public function set($name, $value, $expire = null)
     {
+        $this->writeTimes++;
         if (is_null($expire)) {
             $expire = $this->options['expire'];
         }
@@ -124,6 +126,7 @@ class Memcached extends Driver
      */
     public function inc($name, $step = 1)
     {
+        $this->writeTimes++;
         $key = $this->getCacheKey($name);
         return $this->handler->increment($key, $step);
     }
@@ -137,6 +140,7 @@ class Memcached extends Driver
      */
     public function dec($name, $step = 1)
     {
+        $this->writeTimes++;
         $key   = $this->getCacheKey($name);
         $value = $this->handler->get($key) - $step;
         $res   = $this->handler->set($key, $value);
@@ -155,6 +159,7 @@ class Memcached extends Driver
      */
     public function rm($name, $ttl = false)
     {
+        $this->writeTimes++;
         $key = $this->getCacheKey($name);
         return false === $ttl ?
         $this->handler->delete($key) :
@@ -176,6 +181,7 @@ class Memcached extends Driver
             $this->rm('tag_' . md5($tag));
             return true;
         }
+        $this->writeTimes++;
         return $this->handler->flush();
     }
 }
