+++ library/think/cache/driver/Xcache.php
@@ -61,6 +61,7 @@ class Xcache extends Driver
      */
     public function get($name, $default = false)
     {
+        $this->readTimes++;
         $key = $this->getCacheKey($name);
         return xcache_isset($key) ? xcache_get($key) : $default;
     }
@@ -75,6 +76,7 @@ class Xcache extends Driver
      */
     public function set($name, $value, $expire = null)
     {
+        $this->writeTimes++;
         if (is_null($expire)) {
             $expire = $this->options['expire'];
         }
@@ -98,6 +100,7 @@ class Xcache extends Driver
      */
     public function inc($name, $step = 1)
     {
+        $this->writeTimes++;
         $key = $this->getCacheKey($name);
         return xcache_inc($key, $step);
     }
@@ -111,6 +114,7 @@ class Xcache extends Driver
      */
     public function dec($name, $step = 1)
     {
+        $this->writeTimes++;
         $key = $this->getCacheKey($name);
         return xcache_dec($key, $step);
     }
@@ -123,6 +127,7 @@ class Xcache extends Driver
      */
     public function rm($name)
     {
+        $this->writeTimes++;
         return xcache_unset($this->getCacheKey($name));
     }
 
@@ -143,6 +148,7 @@ class Xcache extends Driver
             $this->rm('tag_' . md5($tag));
             return true;
         }
+        $this->writeTimes++;
         if (function_exists('xcache_unset_by_prefix')) {
             return xcache_unset_by_prefix($this->options['prefix']);
         } else {
