+++ library/think/log/driver/Socket.php
@@ -66,7 +66,7 @@ class Socket
             return false;
         }
         $trace = [];
-        if (App::$debug) {
+        if (App::isDebug()) {
             $runtime    = round(microtime(true) - THINK_START_TIME, 10);
             $reqs       = $runtime > 0 ? number_format(1 / $runtime, 2) : '∞';
             $time_str   = ' [运行时间：' . number_format($runtime, 6) . 's][吞吐率：' . $reqs . 'req/s]';
